+++ changelog.md
@@ -1,3 +1,8 @@
+# 1.7.0
+Add cloudflare-pages-direct-upload command.  
+Usage: `CF_API_TOKEN=token CF_ACCOUNT_ID=id cloudflare-pages-direct-upload PROJECT [DIRECTORY]`  
+If directory is not specified, the current directory is used.
+
 # 1.6.1
 Fix concurrency.